feat: ajust the functions and layout
